requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
scikit-learn
pandas
numpy
